Restructure as monorepo and rebrand to Agent Connector for WP
Turn the repo into a monorepo and complete the rename from "Root for Agents"
to "Agent Connector for WP".

Structure:
- plugin/ — the WordPress plugin (was the repo root). Main file renamed to
  agent-connector-for-wp.php.
- site/  — a one-page static marketing site (HTML/CSS), deployed to GitHub Pages.
- bin/install.sh — symlinks plugin/ into a given wp-content/plugins for local
  dev (requires the wp-content path as an explicit argument; only ever replaces
  an existing symlink, never a real directory).

Rebrand (full):
- namespace RootForAgents\ -> AgentConnectorForWp\
- constants ROOT_FOR_AGENTS_* -> AGENT_CONNECTOR_FOR_WP_*
- ability slugs + category root-for-agents/* -> agent-connector-for-wp/*
- error codes root_for_agents_* -> agent_connector_for_wp_*
- text domain, display name, MCP client name, audit-log filename, and repo URLs
- composer package soflyy/agent-connector-for-wp; PSR-4 autoload regenerated

CI:
- release.yml now builds from the plugin/ subtree and ships
  agent-connector-for-wp.zip.
- pages.yml deploys site/ to GitHub Pages on push to master.
- .gitignore points at plugin/vendor and plugin/dist.

Verified live via a symlinked install: all 9 abilities register under the
agent-connector-for-wp/* slugs, can_boot() is true with the new gate constant,
and an end-to-end MCP call (execute-ability -> agent-connector-for-wp/wp-cli)
returns success with exit_code 0.
