+++ .gitignore
@@ -1,14 +1,14 @@
 # Composer dependencies — installed via `composer install`, bundled into the
 # release zip by CI (see .github/workflows/release.yml). composer.lock IS
 # tracked so builds are reproducible.
-/vendor/
+/plugin/vendor/
 
 # Build artifacts
-/dist/
+/plugin/dist/
 *.zip
 
 # Runtime artifacts
-root-for-agents-audit.log
+agent-connector-for-wp-audit.log
 *.log
 
 # OS / editor cruft
