--- composer.json
@@ -1,25 +1,0 @@
-{
-	"name": "soflyy/root-for-agents",
-	"description": "Give agents unrestricted execution capability on WordPress: shell, PHP eval, filesystem, and process execution exposed through the WordPress MCP stack. Dev only.",
-	"type": "wordpress-plugin",
-	"license": "GPL-2.0-or-later",
-	"keywords": ["wordpress", "mcp", "ai", "agents", "abilities"],
-	"homepage": "https://github.com/soflyy/root-for-agents",
-	"require": {
-		"php": ">=8.1",
-		"automattic/jetpack-autoloader": "^5.0",
-		"wordpress/mcp-adapter": "^0.5"
-	},
-	"autoload": {
-		"psr-4": {
-			"RootForAgents\\": "src/"
-		}
-	},
-	"config": {
-		"allow-plugins": {
-			"automattic/jetpack-autoloader": true
-		},
-		"sort-packages": true
-	},
-	"minimum-stability": "stable"
-}
