--- root-for-agents.php
@@ -1,116 +1,0 @@
-<?php
-/**
- * Plugin Name:       Root for Agents
- * Plugin URI:        https://github.com/soflyy/root-for-agents
- * Description:       Give agents unrestricted execution capability on WordPress. Adds shell access, PHP eval, filesystem operations, and process execution to the WordPress MCP stack. Designed for trusted development environments where agents need the same operational capabilities as a human administrator.
- * Version:           0.1.0
- * Requires at least: 7.0
- * Requires PHP:      8.1
- * Author:            Soflyy
- * Author URI:        https://github.com/soflyy
- * License:           GPL-2.0-or-later
- * License URI:       https://www.gnu.org/licenses/gpl-2.0.html
- * Text Domain:       root-for-agents
- *
- * ---------------------------------------------------------------------------
- *  DANGER: This plugin intentionally grants root-equivalent operational
- *  capability (arbitrary shell, PHP eval, filesystem access) to authenticated
- *  administrators/super admins and the agents acting on their behalf. It is
- *  NOT sandboxed
- *  and NOT intended for production. Enable only in trusted local/dev/staging
- *  environments. See readme.txt.
- * ---------------------------------------------------------------------------
- *
- * @package RootForAgents
- */
-
-declare( strict_types=1 );
-
-namespace RootForAgents;
-
-defined( 'ABSPATH' ) || exit;
-
-define( 'ROOT_FOR_AGENTS_VERSION', '0.1.0' );
-define( 'ROOT_FOR_AGENTS_FILE', __FILE__ );
-define( 'ROOT_FOR_AGENTS_DIR', plugin_dir_path( __FILE__ ) );
-
-/**
- * Load bundled Composer dependencies via the Jetpack Autoloader.
- *
- * We ship wordpress/mcp-adapter (and its php-mcp-schema dependency) inside this
- * plugin so it works standalone — the separate "MCP Adapter" plugin does not
- * need to be installed. The Jetpack Autoloader (autoload_packages.php, not the
- * plain vendor/autoload.php) deduplicates shared packages across plugins: when
- * several plugins bundle the adapter, only the newest version is loaded, which
- * prevents fatal "class already declared" / version-mismatch conflicts.
- *
- * It also registers this plugin's own RootForAgents\ namespace (PSR-4, src/),
- * so the fallback autoloader below is only used when dependencies are missing
- * (e.g. a source checkout where `composer install` has not been run yet).
- */
-$root_for_agents_jetpack_autoloader = ROOT_FOR_AGENTS_DIR . 'vendor/autoload_packages.php';
-$root_for_agents_has_vendor         = is_readable( $root_for_agents_jetpack_autoloader );
-if ( $root_for_agents_has_vendor ) {
-	require_once $root_for_agents_jetpack_autoloader;
-}
-unset( $root_for_agents_jetpack_autoloader );
-
-/**
- * Minimal PSR-4 fallback autoloader for the RootForAgents\ namespace, rooted at
- * src/.
- *
- * Registered only when the Jetpack autoloader is absent (e.g. a source checkout
- * where `composer install` has not been run). When vendor/ is present the
- * Jetpack autoloader already maps the RootForAgents\ namespace, and registering
- * a second loader here would re-require the class files and trigger a "Cannot
- * declare class … already in use" fatal.
- */
-if ( ! $root_for_agents_has_vendor ) {
-	spl_autoload_register(
-		static function ( string $class ): void {
-			$prefix = __NAMESPACE__ . '\\';
-			if ( 0 !== strpos( $class, $prefix ) ) {
-				return;
-			}
-			$relative = substr( $class, strlen( $prefix ) );
-			$path     = ROOT_FOR_AGENTS_DIR . 'src/' . str_replace( '\\', '/', $relative ) . '.php';
-			if ( is_readable( $path ) ) {
-				require $path;
-			}
-		}
-	);
-}
-unset( $root_for_agents_has_vendor );
-
-/**
- * Boot the plugin once WordPress is loaded.
- *
- * The plugin is inert unless the operator has explicitly opted in via the
- * mandatory environment gates (see Support\Config). This keeps a stray
- * activation from ever exposing dangerous abilities by accident.
- */
-add_action(
-	'plugins_loaded',
-	static function (): void {
-		if ( ! Support\Config::can_boot() ) {
-			// Surface *why* in the admin so the gate isn't a silent mystery.
-			add_action( 'admin_notices', array( Support\Config::class, 'render_gate_notice' ) );
-			return;
-		}
-
-		( new Plugin() )->register();
-
-		/**
-		 * Ensure the MCP Adapter is running.
-		 *
-		 * If the standalone "MCP Adapter" plugin is active it will already have
-		 * booted the adapter; McpAdapter::instance() is an idempotent singleton,
-		 * so calling it again is safe. If that plugin is *not* installed, this is
-		 * what brings the bundled adapter (and its default MCP server) to life,
-		 * letting Root for Agents work entirely on its own.
-		 */
-		if ( class_exists( \WP\MCP\Core\McpAdapter::class ) ) {
-			\WP\MCP\Core\McpAdapter::instance();
-		}
-	}
-);
