+++ site/index.html
@@ -1,0 +1,144 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+	<meta charset="UTF-8">
+	<meta name="viewport" content="width=device-width, initial-scale=1.0">
+	<title>Agent Connector for WP — give coding agents real access to WordPress</title>
+	<meta name="description" content="Connect Claude and other coding agents to a WordPress site over MCP, with shell, WP-CLI, PHP eval, and filesystem access. For trusted local/dev environments.">
+	<link rel="stylesheet" href="style.css">
+</head>
+<body>
+	<header class="hero">
+		<div class="container">
+			<p class="eyebrow">WordPress &times; MCP</p>
+			<h1>Agent Connector for WP</h1>
+			<p class="tagline">
+				Connect Claude — or any MCP-capable coding agent — to a WordPress site
+				and give it real operational access: shell, WP-CLI, PHP eval, and the
+				filesystem.
+			</p>
+			<div class="cta-row">
+				<a class="btn btn-primary" href="https://github.com/soflyy/agent-connector-for-wp">View on GitHub</a>
+				<a class="btn btn-ghost" href="#install">Install</a>
+			</div>
+			<p class="danger-pill">⚠️ Dangerous by design — trusted dev/staging only, never production.</p>
+		</div>
+	</header>
+
+	<main>
+		<section class="section">
+			<div class="container">
+				<h2>What it does</h2>
+				<p class="lead">
+					WordPress' MCP stack exposes structured abilities, but agents still
+					can't <em>operate</em> the server. Agent Connector for WP closes that
+					gap: it registers operational abilities and surfaces them over MCP,
+					then hands you a one-click way to connect an agent to them.
+				</p>
+			</div>
+		</section>
+
+		<section class="section section-alt">
+			<div class="container">
+				<h2>Abilities</h2>
+				<div class="grid">
+					<div class="card">
+						<h3>Shell</h3>
+						<p>Run arbitrary shell commands, capturing stdout, stderr, exit code — with cwd and timeout.</p>
+					</div>
+					<div class="card">
+						<h3>WP-CLI</h3>
+						<p>Run any <code>wp</code> command against the install; <code>--allow-root</code> is added automatically when needed.</p>
+					</div>
+					<div class="card">
+						<h3>PHP eval</h3>
+						<p>Evaluate PHP inside the live WordPress runtime; get back output, the return value, and errors.</p>
+					</div>
+					<div class="card">
+						<h3>Filesystem</h3>
+						<p>Read, write, delete, and list files — binary-safe via base64.</p>
+					</div>
+					<div class="card">
+						<h3>Environment</h3>
+						<p>Inspect versions, paths, active plugins/theme, debug flags, and available CLI tooling.</p>
+					</div>
+					<div class="card">
+						<h3>Process</h3>
+						<p>Longer-running command execution, the forward-looking home for async/streaming control.</p>
+					</div>
+				</div>
+			</div>
+		</section>
+
+		<section class="section" id="connect">
+			<div class="container">
+				<h2>One-click connect</h2>
+				<p class="lead">
+					A <strong>Connect</strong> admin page generates everything an agent
+					needs in one click: it mints a WordPress application password,
+					computes the MCP endpoint, and hands you three copy-paste artifacts.
+				</p>
+				<ul class="checks">
+					<li>A natural-language prompt to paste into any coding agent</li>
+					<li>A <code>claude mcp add</code> command for Claude Code</li>
+					<li>An <code>mcpServers</code> JSON block for client config files</li>
+				</ul>
+				<p class="note">
+					All three drive the
+					<a href="https://www.npmjs.com/package/@automattic/mcp-wordpress-remote">@automattic/mcp-wordpress-remote</a>
+					proxy and authenticate with the application password.
+				</p>
+			</div>
+		</section>
+
+		<section class="section section-alt" id="install">
+			<div class="container">
+				<h2>Install</h2>
+				<p class="lead">It's a monorepo — the plugin lives in <code>plugin/</code>. For local development, symlink it into your WordPress install:</p>
+				<pre><code>git clone https://github.com/soflyy/agent-connector-for-wp.git
+cd agent-connector-for-wp
+
+# Install the plugin's PHP dependencies
+( cd plugin &amp;&amp; composer install --no-dev )
+
+# Symlink plugin/ into wp-content/plugins/agent-connector-for-wp
+bin/install.sh /path/to/wp-content
+
+wp plugin activate agent-connector-for-wp</code></pre>
+				<p class="note">
+					Then enable it in <code>wp-config.php</code> with
+					<code>define( 'AGENT_CONNECTOR_FOR_WP_ENABLED', true );</code>. The
+					plugin refuses to run on production unless explicitly overridden.
+					Prefer a packaged release <code>.zip</code>? It already bundles
+					dependencies — just activate.
+				</p>
+			</div>
+		</section>
+
+		<section class="section danger">
+			<div class="container">
+				<h2>Read this first</h2>
+				<p>
+					This plugin is <strong>intentionally unrestricted</strong>. Anyone
+					holding a generated application password can run shell commands,
+					evaluate PHP, and read/write files on the server as you. There is no
+					sandboxing, no command allowlist, no approval workflow. Install it
+					only where you would be comfortable handing out a root shell, and
+					revoke application passwords when you're done.
+				</p>
+			</div>
+		</section>
+	</main>
+
+	<footer class="footer">
+		<div class="container">
+			<p>
+				<a href="https://github.com/soflyy/agent-connector-for-wp">GitHub</a> ·
+				<a href="https://github.com/soflyy/agent-connector-for-wp/blob/master/plugin/README.md">Docs</a> ·
+				Built by <a href="https://github.com/soflyy">Soflyy</a>
+			</p>
+			<p class="muted">GPL-2.0-or-later</p>
+		</div>
+	</footer>
+</body>
+</html>
