+++ site/style.css
@@ -1,0 +1,176 @@
+:root {
+	--bg: #0f1117;
+	--bg-alt: #161922;
+	--surface: #1d212c;
+	--border: #2a2f3d;
+	--text: #e6e8ee;
+	--muted: #9aa1b1;
+	--accent: #6e8bff;
+	--accent-2: #9b6bff;
+	--danger: #ff6b6b;
+	--danger-bg: #2a1416;
+	--radius: 12px;
+	--maxw: 960px;
+}
+
+* { box-sizing: border-box; }
+
+html { scroll-behavior: smooth; }
+
+body {
+	margin: 0;
+	font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Helvetica, Arial, sans-serif;
+	background: var(--bg);
+	color: var(--text);
+	line-height: 1.6;
+	-webkit-font-smoothing: antialiased;
+}
+
+.container {
+	width: 100%;
+	max-width: var(--maxw);
+	margin: 0 auto;
+	padding: 0 24px;
+}
+
+a { color: var(--accent); text-decoration: none; }
+a:hover { text-decoration: underline; }
+
+code {
+	font-family: ui-monospace, SFMono-Regular, Menlo, Consolas, monospace;
+	font-size: 0.9em;
+	background: rgba(255, 255, 255, 0.07);
+	padding: 0.15em 0.4em;
+	border-radius: 5px;
+}
+
+/* Hero */
+.hero {
+	padding: 96px 0 72px;
+	text-align: center;
+	background:
+		radial-gradient(1100px 480px at 50% -10%, rgba(110, 139, 255, 0.18), transparent 70%),
+		var(--bg);
+	border-bottom: 1px solid var(--border);
+}
+
+.eyebrow {
+	text-transform: uppercase;
+	letter-spacing: 0.18em;
+	font-size: 0.8rem;
+	color: var(--muted);
+	margin: 0 0 12px;
+}
+
+.hero h1 {
+	font-size: clamp(2.4rem, 6vw, 3.8rem);
+	line-height: 1.05;
+	margin: 0 0 16px;
+	background: linear-gradient(90deg, var(--accent), var(--accent-2));
+	-webkit-background-clip: text;
+	background-clip: text;
+	-webkit-text-fill-color: transparent;
+}
+
+.tagline {
+	font-size: 1.2rem;
+	color: var(--muted);
+	max-width: 640px;
+	margin: 0 auto 28px;
+}
+
+.cta-row {
+	display: flex;
+	gap: 12px;
+	justify-content: center;
+	flex-wrap: wrap;
+	margin-bottom: 28px;
+}
+
+.btn {
+	display: inline-block;
+	padding: 12px 22px;
+	border-radius: 8px;
+	font-weight: 600;
+	border: 1px solid transparent;
+	transition: transform 0.06s ease, opacity 0.2s ease;
+}
+.btn:hover { text-decoration: none; transform: translateY(-1px); }
+.btn-primary { background: linear-gradient(90deg, var(--accent), var(--accent-2)); color: #fff; }
+.btn-ghost { border-color: var(--border); color: var(--text); }
+
+.danger-pill {
+	display: inline-block;
+	margin: 0;
+	font-size: 0.9rem;
+	color: var(--danger);
+	background: var(--danger-bg);
+	border: 1px solid rgba(255, 107, 107, 0.35);
+	padding: 8px 16px;
+	border-radius: 999px;
+}
+
+/* Sections */
+.section { padding: 72px 0; }
+.section-alt { background: var(--bg-alt); border-top: 1px solid var(--border); border-bottom: 1px solid var(--border); }
+
+h2 { font-size: clamp(1.6rem, 3.5vw, 2.2rem); margin: 0 0 20px; }
+h3 { margin: 0 0 8px; font-size: 1.15rem; }
+
+.lead { font-size: 1.1rem; color: var(--muted); max-width: 720px; }
+.lead em { color: var(--text); font-style: italic; }
+
+/* Cards */
+.grid {
+	display: grid;
+	grid-template-columns: repeat(auto-fit, minmax(240px, 1fr));
+	gap: 16px;
+	margin-top: 28px;
+}
+
+.card {
+	background: var(--surface);
+	border: 1px solid var(--border);
+	border-radius: var(--radius);
+	padding: 22px;
+}
+.card p { margin: 0; color: var(--muted); }
+
+/* Lists */
+.checks { list-style: none; padding: 0; margin: 24px 0 0; max-width: 720px; }
+.checks li { padding: 10px 0 10px 32px; position: relative; border-bottom: 1px solid var(--border); }
+.checks li::before {
+	content: "✓";
+	position: absolute;
+	left: 0;
+	color: var(--accent);
+	font-weight: 700;
+}
+
+.note { color: var(--muted); font-size: 0.95rem; margin-top: 20px; }
+
+/* Code block */
+pre {
+	background: #0b0d13;
+	border: 1px solid var(--border);
+	border-radius: var(--radius);
+	padding: 20px;
+	overflow-x: auto;
+	margin-top: 24px;
+}
+pre code { background: none; padding: 0; font-size: 0.88rem; line-height: 1.7; color: #cdd3e1; }
+
+/* Danger section */
+.danger { background: var(--danger-bg); border-top: 1px solid rgba(255, 107, 107, 0.25); }
+.danger h2 { color: var(--danger); }
+.danger p { color: #f0c9c9; max-width: 760px; }
+
+/* Footer */
+.footer {
+	padding: 40px 0;
+	text-align: center;
+	border-top: 1px solid var(--border);
+	color: var(--muted);
+}
+.footer p { margin: 4px 0; }
+.muted { color: var(--muted); font-size: 0.85rem; }
